pyproject.toml
[tool.poetry]
name = "cadeia_abastecimento"
version = "0.1.0"
description = ""
authors = ["Iago Barbosa <iago@estudante.ufscar.br>"]

[tool.poetry.dependencies]
python = "^3.8"
pytest = "^7.1.3"
paho-mqtt = "1.5.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
